Action button on all contexts.
Combined the registration status and id into one LabelField: status

Alert and ManDown have a countdown timer.

Issue: the status text does not change when a message is sent. This
is because there is no delay in changing the status text. Putting
in a Thread.sleep won't work as you can't sleep the event lock.
So this probably needs to be called from outside.