Bug fix: duplicate lockscreen getting pushed.
Added a check to make sure that when the volume up is double tapped that
an additional lockscreen is not pushed when the active screen is the
lockscreen itself.

To accomplish this, the lockscreen had to be instantiated in the
LockKeyListener constructor. This makes the app a bit more memory
intensive and may delay initial startup time. Another solution would be to
store a flag in the runtime store that updates whenever the screen is
pushed or popped.