Changed event lock access for LockScreen push
Was previously synchronizing on the event lock but changed to the push
request to the event queue. The reason being that the event queue should
only be held for a very short period of time and since the pushing of the
screen could take longer it is safer to do an invoke and wait.